Mutated dataframes and added graphs
Mutated the data to find the reviews per year. Also improved the Treemaps by changing the area to be based on market size (available days per price). Also created a new Treemap that compares the market size by neighbourhood and has a heatmap that shows which neighbourhoods have the largest ratio of reviews per listing days to determine which neighbourhoods have the largest disjointement between supply and demand and thus more potential opportunity for investment.